+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_spm_half_cell.py
@@ -1,22 +1,12 @@
 #
 # Tests for the lithium-ion half-cell SPM model
 #
-from tests import TestCase
 import pybamm
-import unittest
 from tests import BaseUnitTestLithiumIonHalfCell
+import pytest
 
 
-class TestSPMHalfCell(BaseUnitTestLithiumIonHalfCell, TestCase):
+class TestSPMHalfCell(BaseUnitTestLithiumIonHalfCell):
+    @pytest.fixture(autouse=True)
     def setUp(self):
         self.model = pybamm.lithium_ion.SPM
-
-
-if __name__ == "__main__":
-    print("Add -v for more debug output")
-    import sys
-
-    if "-v" in sys.argv:
-        debug = True
-    pybamm.settings.debug_mode = True
-    unittest.main()